latest

+++ ndn_prefix/runProcessGeocodeJson.sh
@@ -11,7 +11,11 @@ mv testbedNodes.json testbedNodes.json.PREV
 mv processGeocodeJson.log  processGeocodeJson.log.PREV
 ./processGeocodeJson.py geocode.json testbedNodes.json >& processGeocodeJson.log
 chmod 644 testbedNodes.json
-sudo cp testbedNodes.json /var/www/html/
-sudo cp links.json /var/www/html/
-sudo rm /var/www/html/testbed-nodes.json
-sudo ln -s /var/www/html/testbedNodes.json /var/www/html/testbed-nodes.json
+if [ -s testbedNodes.json ]
+then
+  sudo cp testbedNodes.json /var/www/html/
+  sudo cp links.json /var/www/html/
+  sudo rm /var/www/html/testbed-nodes.json
+  sudo ln -s /var/www/html/testbedNodes.json /var/www/html/testbed-nodes.json
+fi
+
